requirements.txt
absl-py==0.12.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
config==0.5.0.post0
et-xmlfile==1.1.0
cycler==0.10.0
enum34==1.1.10
et-xmlfile==1.1.0
fastapi==0.67.0
filelock==3.0.12
future==0.18.2
gin-config==0.1.4
googleapis-common-protos==1.53.0
h11==0.12.0
huggingface-hub==0.0.12
idna==3.2
joblib==1.0.1
numpy==1.21.1
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
openai==0.10.2
openpyxl==3.0.7
packaging==21.0
pandas==1.3.1
pandas-stubs==1.2.0.1
Pillow==8.3.1
promise==2.3
protobuf==3.17.3
pstar==0.1.9
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
sacremoses==0.0.45
starlette==0.14.2
qj==0.2.0
ray==1.11.0
requests==2.26.0
scipy==1.7.0
six==1.16.0
tensorflow-datasets==1.0.1
tensorflow-gan==0.0.0.dev0
tensorflow-hub==0.12.0
tensorflow-metadata==1.2.0
termcolor==1.1.0
tokenizers==0.10.3
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
transformers==4.9.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wrapt==1.12.1
